Unify modules and add graph export
